Begin changing i32 values to usize

+++ src/xyzzy.rs
@@ -74,8 +74,8 @@ impl WinMineMainWindow {
         // Check if the Control key is held down.
         let control_down = key == MK::CONTROL;
         if (state == CCH_XYZZY && control_down) || state > CCH_XYZZY {
-            let x_pos = self.x_box_from_xpos(point.x);
-            let y_pos = self.y_box_from_ypos(point.y);
+            let x_pos = self.x_box_from_xpos(point.x) as i32;
+            let y_pos = self.y_box_from_ypos(point.y) as i32;
             self.state.write().cursor_pos = POINT { x: x_pos, y: y_pos };
             // Check if the cursor is within the board's range
             let in_range = x_pos > 0
